[DEP-103] rename test-data folder back to testData, remove test-examples folder with the demo tests
