+++ client/e2e/testData/dataConstants/CustomerInfoConstants.ts
