+++ client/e2e/testData/dataConstants/EventInfoConstants.ts
